Adding mxnet task
