+++ tasks/__init__.py
@@ -1,0 +1,14 @@
+from invoke import Collection
+
+from . import (
+    cpython,
+    libs,
+    mxnet,
+)
+
+ns = Collection(
+    cpython,
+    libs,
+    mxnet,
+)
+
